refactor: share faucet verification overlay
