+++ idea/gear/frontend/src/features/balance/ui/verification-overlay/index.ts
@@ -1,0 +1,1 @@
+export { VerificationOverlay } from './verification-overlay';
