+++ idea/gear/frontend/src/features/balance/ui/verification-overlay/verification-overlay.module.scss
